Move seat-related endpoints to correct Controllers.
